Inject and use existing ZooKeeper
