+++ controller-server/src/main/java/com/yahoo/vespa/hosted/controller/persistence/CuratorDb.java
@@ -2,10 +2,7 @@
 package com.yahoo.vespa.hosted.controller.persistence;
 
 import com.google.inject.Inject;
-import com.yahoo.cloud.config.ClusterInfoConfig;
-import com.yahoo.cloud.config.ZookeeperServerConfig;
 import com.yahoo.config.provision.ApplicationId;
-import com.yahoo.net.HostName;
 import com.yahoo.path.Path;
 import com.yahoo.transaction.NestedTransaction;
 import com.yahoo.vespa.config.SlimeUtils;
@@ -14,7 +11,6 @@ import com.yahoo.vespa.curator.Lock;
 import com.yahoo.vespa.hosted.controller.api.identifiers.TenantId;
 import com.yahoo.vespa.hosted.controller.application.DeploymentJobs;
 import com.yahoo.vespa.hosted.controller.versions.VersionStatus;
-import com.yahoo.vespa.zookeeper.ZooKeeperServer;
 
 import java.io.IOException;
 import java.io.UncheckedIOException;
@@ -30,7 +26,6 @@ import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import java.util.stream.Collectors;
 
 /**
  * Curator backed database for storing working state shared between controller servers.
@@ -39,9 +34,6 @@ import java.util.stream.Collectors;
  * @author bratseth
  */
 public class CuratorDb {
-
-    /** Use a nonstandard zk port to avoid interfering with connection to the config server zk cluster */
-    private static final int zooKeeperPort = 2281;
 
     private static final Logger log = Logger.getLogger(CuratorDb.class.getName());
 
@@ -52,9 +44,6 @@ public class CuratorDb {
     private final StringSetSerializer stringSetSerializer = new StringSetSerializer();
     private final JobQueueSerializer jobQueueSerializer = new JobQueueSerializer();
 
-    @SuppressWarnings("unused") // This server is used (only) from the curator instance of this over the network */
-    private final ZooKeeperServer zooKeeperServer;
-
     private final Curator curator;
 
     /**
@@ -63,52 +52,9 @@ public class CuratorDb {
      */
     private final ConcurrentHashMap<Path, Lock> locks = new ConcurrentHashMap<>();
 
-    /** Create a curator db which also set up a ZooKeeper server (such that this instance is both client and server) */
     @Inject
-    public CuratorDb(ClusterInfoConfig clusterInfo) {
-        this.zooKeeperServer = new ZooKeeperServer(toZookeeperServerConfig(clusterInfo));
-        this.curator = new Curator(toConnectionSpec(clusterInfo));
-    }
-
-    /** Create a curator db which does not set up a server, using the given Curator instance */
-    protected CuratorDb(Curator curator) {
-        this.zooKeeperServer = null;
+    public CuratorDb(Curator curator) {
         this.curator = curator;
-    }
-
-    private static ZookeeperServerConfig toZookeeperServerConfig(ClusterInfoConfig clusterInfo) {
-        ZookeeperServerConfig.Builder b = new ZookeeperServerConfig.Builder();
-        b.zooKeeperConfigFile("conf/zookeeper/controller-zookeeper.cfg");
-        b.dataDir("var/controller-zookeeper");
-        b.clientPort(zooKeeperPort);
-        b.myidFile("var/controller-zookeeper/myid");
-        b.myid(myIndex(clusterInfo));
-
-        for (ClusterInfoConfig.Services clusterMember : clusterInfo.services()) {
-            ZookeeperServerConfig.Server.Builder server = new ZookeeperServerConfig.Server.Builder();
-            server.id(clusterMember.index());
-            server.hostname(clusterMember.hostname());
-            server.quorumPort(zooKeeperPort + 1);
-            server.electionPort(zooKeeperPort + 2);
-            b.server(server);
-        }
-        return new ZookeeperServerConfig(b);
-    }
-
-    private static Integer myIndex(ClusterInfoConfig clusterInfo) {
-        String hostname = HostName.getLocalhost();
-        return clusterInfo.services().stream()
-                .filter(service -> service.hostname().equals(hostname))
-                .map(ClusterInfoConfig.Services::index)
-                .findFirst()
-                .orElseThrow(() -> new IllegalStateException("Unable to find index for this node by hostname '" +
-                                                             hostname + "'"));
-    }
-
-    private static String toConnectionSpec(ClusterInfoConfig clusterInfo) {
-        return clusterInfo.services().stream()
-                .map(member -> member.hostname() + ":" + zooKeeperPort)
-                .collect(Collectors.joining(","));
     }
 
     // -------------- Locks --------------------------------------------------
